move components to app dir
