+++ src/app/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import '../assets/styles/App.css';
-import Nav from "../components/Nav/Nav";
+import Nav from "./components/Nav/Nav";
 import {Outlet} from "react-router-dom";
 
 function App() {
