+++ src/app/components/Nav/Nav.tsx
@@ -17,8 +17,8 @@ function Nav() {
                 <Link to="/blog">Blog</Link>
             </div>
             <div className="nav-section nav-btn-grp">
-                <a className="nav-button" href="#">Resume <VscFilePdf className="nav-icon"/></a>
-                <a className="nav-button" href="#">Book <AiOutlineCalendar className="nav-icon" /></a>
+                <a className="nav-button" href="app/components/Nav/Nav#">Resume <VscFilePdf className="nav-icon"/></a>
+                <a className="nav-button" href="app/components/Nav/Nav#">Book <AiOutlineCalendar className="nav-icon" /></a>
                 <Link to="/budget" className="nav-button">Budget</Link>
             </div>
         </nav>
