+++ src/app/pages/Home/Home.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import CodeBox from "../../../components/CodeBox/CodeBox";
+import CodeBox from "../../components/CodeBox/CodeBox";
 import AppConfig from "../../../config";
 
 function Home() {
